examples/requirements.txt
alembic==1.4.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cloudml-hypertune==0.1.0.dev6
databricks-cli==0.14.3
docker==5.0.0
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.26.3
google-api-python-client==2.18.0
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
jupyter==1.0.0
Mako==1.1.4
MarkupSafe==1.1.1
mlflow==1.16.0
numpy==1.20.2
oauth2client==4.1.3
packaging==20.9
pandas==1.2.4
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
protobuf==3.15.8
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
tabulate==0.8.9
typing-extensions==3.10.0.0
urllib3==1.26.5
websocket-client==0.59.0
Werkzeug==1.0.1
zipp==3.4.1

